* Add `implements<Interface, &Actor::methods...>` for interface-based polymorphism
* Add custom `promise_type` support per actor via `Actor::promise_type<T>`

* Add type erasure: `unique_future<T>` → `unique_future<void>`

* Add `from_state()` factory for custom promise implementations

* Add `make_ready_future<T>()` helpers

* Add `is_interface` / `is_actor` concepts

* Add `send()` overload for `address_t` with interface polymorphism

* Fix GCC `-Wmaybe-uninitialized` false positive in future_state

* Fix `-Wshadow` warnings with stricter compiler flags

* Add stricter compiler warning flags

+++ test/dispatcher-pattern/manager_dispatcher.hpp
@@ -570,7 +570,7 @@ public:
             }
             g_log.log("[%::process_batch_buffer] Processing % items", name_, b->size());
             result.committed = true;
-            result.total_rows = static_cast<int>(b->size());
+            result.total_rows = b->size();
             co_return result;
         };
         co_return co_await process(resource());
